if you must auto format then at least do not corrupt the file
It is rarely a good idea to do any auto format as part of the build, and
less so to do it using a tool that is gaunranteed to corrupt the file.

The formatter is borked and would not read and write the file with the
correct encoding and would hence garble the encoding.

this would lead to the following when running `mvn validate` twice

```
[INFO] --- xml-format-maven-plugin:3.0.7:xml-format (validate) @
google-storage-plugin ---
[ERROR] [xml-format] Error for
C:\workarea\source\github\jenkinsci\google-storage-plugin\pom.xml
org.dom4j.DocumentException: Invalid byte 3 of 3-byte UTF-8 sequence.
    at org.dom4j.io.SAXReader.read (SAXReader.java:468)
    at org.dom4j.io.SAXReader.read (SAXReader.java:329)
    at au.com.acegi.xmlformat.FormatUtil.format (FormatUtil.java:78)
    at au.com.acegi.xmlformat.FormatUtil.formatInPlace (FormatUtil.java:107)
```